backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.12

boto3==1.34.162
pillow==10.4.0
